fix(filter-field): Fixes an issue with async data and programmatically set filters.
When setting the filters programmatically, as the quickfilter does when a filter is set, the
retreival of set filters breaks.
The programmatic setting of the filters runs a function called `tryApplyFilters`, which looks in the
datasource and async dataset for the correct filters.
For some reason, the root of the async def was pushed twice, resulting in a wrong path and a wrong
id.

Fixes APM-267380